Move pkg/server/grpc to pkg/server/ca.
