+++ security/pkg/server/ca/server.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package grpc
+package ca
 
 import (
 	"crypto/tls"
